refactor(core): 重构了部分文件(#246)
